Update convert.py
try this update -- I think this might work as it would appear that Jenkins is picking up other notebooks that are being installed in the path but not in the repository.    I'm not 100% this will work, but I believe updating this to only look for notebooks in this repository will remove the problematic notebook.